turn recipeId back into int for review fragment

+++ app/src/main/java/com/example/wat2eat/ui/recipe/RecipeFragment.kt
@@ -53,7 +53,7 @@ class RecipeFragment : Fragment() {
         recipeViewModel = ViewModelProvider(this)[RecipeViewModel::class.java]
         binding.seeReviews.setOnClickListener {
             val intent = Intent(context, ReviewActivity::class.java).apply {
-                putExtra(ReviewActivity.EXTRA_RECIPE_ID, recipeId)
+                putExtra(ReviewActivity.EXTRA_RECIPE_ID, recipeId.toInt())
             }
             startActivity(intent)
         }
